requirements.txt
python-telegram-bot==20.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
loguru==0.7.0
telethon==1.40.0
asyncio==3.4.3 